+++ src/XUnity.AutoTranslator.Setup/Properties/Resources.Designer.cs
@@ -19,7 +19,7 @@ namespace XUnity.AutoTranslator.Setup.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "15.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -146,6 +146,26 @@ namespace XUnity.AutoTranslator.Setup.Properties {
         internal static byte[] GoogleTranslate {
             get {
                 object obj = ResourceManager.GetObject("GoogleTranslate", resourceCulture);
+                return ((byte[])(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Byte[].
+        /// </summary>
+        internal static byte[] GoogleTranslateCompat {
+            get {
+                object obj = ResourceManager.GetObject("GoogleTranslateCompat", resourceCulture);
+                return ((byte[])(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Byte[].
+        /// </summary>
+        internal static byte[] GoogleTranslateCompat_ExtProtocol {
+            get {
+                object obj = ResourceManager.GetObject("GoogleTranslateCompat_ExtProtocol", resourceCulture);
                 return ((byte[])(obj));
             }
         }